services and DAOs are ready (no DAOs for all Entities yet)

+++ src/main/java/com/pacard/dao/DiagnosisDao.java
@@ -19,7 +19,7 @@ public class DiagnosisDao {
         em.persist(diagnosis);
     }
 
-    public Diagnosis findById(int id) {
+    public Diagnosis findById(Long id) {
         return em.find(Diagnosis.class, id);
     }
